modified binary query from base64 to hex
