+++ src/table/persistence/ICollumnQueryType.ts
@@ -1,0 +1,6 @@
+export interface ICollumnQueryType {
+  isGuid(): boolean;
+  isBinary(): boolean;
+  isOther(): boolean;
+  getPrefix(): string;
+}
