+++ src/table/persistence/OtherCollumnQueryType.ts
@@ -1,0 +1,16 @@
+import { ICollumnQueryType } from "./ICollumnQueryType";
+
+export class OtherCollumnQueryType implements ICollumnQueryType {
+  isGuid(): boolean {
+    return false;
+  }
+  isBinary(): boolean {
+    return false;
+  }
+  isOther(): boolean {
+    return true;
+  }
+  getPrefix(): string {
+    return "";
+  }
+}
